• 🦥 feat (watchmovie.ts): add check for available tickets on sale
Add a new prop to the WatchMovie component that allows it to check whether tickets are on sale for the movie. If tickets are on sale, the watch text will be disabled and the text will change to "Смотреть в кино". This change improves the user experience by preventing users from watching movie that are not available.

Changes made:
- Add new prop "ticketsOnSale" to WatchMovie component
- Update button text and disable button if tickets are on sale
- Update styles for disabled button
- Updated interface for props